Move Objective-C.gitignore to Global/Xcode.gitignore
The rules which were previously located in Objective-C.gitignore are
purely Xcode related, and therefore should be located in
Global/Xcode.gitignore, as are the rules for every other IDE.

This comes in handy because one may be working with Xcode in another
language (e.g. C/C++) and would probably not look in the Objective-C.gitignore
file for the rules.

Then again, I have to admit that most people working with Objective-C
are in fact working with Xcode, which is why I created an alias to the
Xcode rules with the old name, that is, Objective-C.gitignore