Ignore LaTeX generated files of extension .fls.
From latexmk_4.35-0ubuntu1_all manpage about the -recorder option:

Use the -recorder option with latex and pdflatex. In (most) modern
versions of these programs, this results in a file of extension .fls
containing a list of the files that these programs have read and
written. Latexmk will then use this file to improve its detection of
source files and generated files after a run of latex or pdflatex.