Added the new upload directories and ignore the cache directory completely.
With SugarCRM 6.4 the `cache/upload` directory moved to `upload` and
`upload_backup`. These directories can safely be ignored.

Additionally the `cache` directory is now ignored completely because it
leads to the behaviour to use the git repository as a kind of backup
tool if several parts of it are included. Furthermore the directory
structure beneath `cache` is victim of enduring changes by the
developers.

I did also some code cleanup regarding the path names e.g. adding a `/`
before most file and directory names.