add Meta Programming System gitignore files
The Meta Programming System is a language workbench created by JetBrains.

http://jetbrains.com/mps/

By default this file will ignore the generated java source and the compiled java classes that are
produced by the language compiler. It will only add the Models which contain all information
necessary to reproduce the the java source code and the the resulting class files.

In addition to the generated source code it will ignore the test result files and test property files,
because those are also generated from the models.

The ignore file also prevents commiting the workspace.xml which contains the currently editor state,
like open files and cursor position.