Keep Visual Studio and XCode project files
The original author excluded XCode .xcodeproj files and Visual Studio .sln files. Personally, in my current project, I would prefer to include the (lightweight) project files in my repo because it allows anybody to clone the repo and then immediately build the project from the IDE. Without the files, the cloner has to open the .uproject file in UE4 Editor and go to File > Generate [Visual Studio, XCode] Project. Sometimes, the user may want to just open the project in the IDE and build the Release version without having to fiddle with the Editor.

This is of course a personal preference/opinion but I believe it would be worth implementing.
If so, lines 37 and 38 should also be noted as being important because XCode stores a bunch of user-specific garbage inside the .xcodeproj folder.