Update 'Icon\r\r' rule in macOS.gitingore
First things first, I haven't tested this, so there's a chance I'm wrong about it being a solution. I'm not the contributor to my project with a Mac, so I cannot test it myself.

I merged this file into my .gitignore file, using Sublime Text, which displayed the \r character as [CR] as shown here: http://i.imgur.com/bXmRxoe.png
Saving the file and re-opening it removes those markers and replaces it with new lines, and I'm not certain if these new lines are still \r or not. I had a quick google, and came across this StackOverflow answer: http://i.imgur.com/bXmRxoe.png
It appears, according to that answer, that my change does the same thing as the current rule without the chance of being converted to a normal new line marker, which is dependent on the operating system. There's also a chance that this new rule is specific for 'Icon\r', but not knowing anything about macOS, I can't be sure that it's valid since the current rule is for 'Icon\r\r', which means an additional '[\r]' or '?' may be necessary.

I'm putting this here in the hope that someone can test it on my behalf, which would either lead to slightly better cross-platform support for this particular rule, or to me being set on the right path for adding this to my .gitignore file.