Add *.snk strong name key files
Strong name key files shouldn't be included in a repository AFAIK. they are intended to sign build output to verify that it comes from the correct publisher. having an *.snk in a repository would allow anyone to produce build as if they were the original publisher.

I guess some OSS projects might like to have *.snk files in their repos but that would be an exception to the rule.