Update Unity.gitignore
"Unity3D is a powerful cross-platform 3D engine and a user friendly development environment. Easy enough for the beginner and powerful enough for the expert;" 

"user friendly",  "beginner" case in point; having files not be committed because they have no knowledge that they are not supposed to name their folders  "library" or "build"  inside unity is not cool. 

The point of a default should be to only care about the important case, ie the magic folders that Unity creates and leave all other folders alone.