Do not ignore launchsettings.json for VS
Ignoring `launchSettings.json` does not make much sense. Now .NET CLI even considers this file when running with `dotnet run`, as you can read [here](https://docs.microsoft.com/en-us/dotnet/core/tools/dotnet-run?tabs=netcore2x).

This settings will be useful if shared among project members, so it should be commited to the repo.

Also, on the default `.gitignore` file generated by Visual Studio it is not ignored, so this causes confusion, as depending on how `.gitignore` was created it could be commited in or not.