Added Rider (IDEA) support
Adding JetBrains Rider support since the reason why it was removed it false. There is no ignore file for Rider. Also since Rider supports .NET and .NET Core why should it not be included in this ignore file. Would just replicate almost all of the entries present in this ignore file.