Update Android.gitignore
Updated google-services.json rule to have a wildcard prefix and also added the new google_maps_api file as it contains the development/production key when using maps API.

**Reasons for making this change:**
Uncommenting the rule doesn't exclude it on account of it being inside the app/ folder than the root directory (as clearly mentioned in the google documentation as well). So the wildcard has been appended. The google_maps_api rule is self explanatory.

**Links to documentation supporting these rule changes:**
https://developers.google.com/android/guides/google-services-plugin#adding_the_json_file
https://developers.google.com/maps/documentation/android-sdk/get-api-key#getting-the-certification-information-from-android-studio