Adding TeXstudio.gitignore
This gitignore can be useful for LaTeX XeTeX, LuaTex in general and TeXstudio projects in general.
*TeX creates a plethora of temporary files which will be re-generated when compiling the .tex, bibilography and style files. All these files should not be included in versioning systems like git.
You may wonder why .pdf files are not ignored, as the compiled result of a TeX project is a pdf. This is because illustrations in a TeX document can be pdf files which you will need for compilation.