Add JetBrains note for PyCharm to Python.gitignore
First and foremost I think that requiring such a complicated gitignore for JetBrains IDEs is a failure on JetBrains part in structuring their project setting.  I also feel that it should be included in the `Python.gitignore` due to its popularity and due to the frequency of requests.  A quick search for `PyCharm` PRs shows 81 closed PRs requesting it be added and if searching for `.idea` you can see many more have been requested.  However I understand the maintenance burden in including it when a user can manually merge the two files themselves so I understand why the maintainers have opted to maintain it seperately.  

The main problem I see is that with many people adding the `Python.gitignore` at project creation through the Github UI and the `JetBrains.gitignore` being in the Global folder and makes it less discoverable than it should be.  

This PR adds a comment for people adding the `Python.gitignore` directing them to the global `JetBrains.gitignore` which should solve this issue in a way that is acceptable for the maintainers, since comment-only sections already exist for `pyenv` and `pipenv`.