Added .vscode  in python.gitignore
While using vs code when choosing an interpreter for python, a ".vscode" file is generated and I notice that this is not included in the gitignore file. But I'm recommending to add this to the list. There might be related cache files for other editors, but I just use vs code.