Unity: Remove exclusion for *.meta files
If a user has ignored an asset (no matter how dubious that decision may be), they also want to ignore the .meta file. Instead of bringing back .metas that have been ignored, this template should trust that a user has ignored the files they want to ignore.

You'd encounter this issue if you had added an ignore for an asset and its meta above this Unity template, or if another template ignored an asset and meta.

In general excludes may be dangerous in these templates, as they can have unintended consequences on other templates.