Ignore Rider files
Added ignore statement for .idea/ and the RiderLink plugin.

Link to Rider for Unreal Engine: https://www.jetbrains.com/lp/rider-unreal/

Rider automatically creates the .idea/ folder which contains settings for the IDE. These files are not supposed to be pushed to the repository since everyone uses a different IDE to program in Unreal Engine.

For complete support between Rider and Unreal Engine, Rider requires you to install the RiderLink plugin. This plugin contains 500+ files. These files are not supposed to be pushed to the repository since everyone uses a different IDE to program in Unreal Engine.