Call CC0 copying information instead of license
Before this change, this repo said that its "license" was CC0. CC0 is a
public domain dedication, not a license. (See
<https://creativecommons.org/faq/#how-do-cc-licenses-operate>). This
change makes this repo refer to CC0 as copying information.

The term "copying" was chosen to make this repo more closely follow CC's
recommendation for applying CC0 to computer software:
<https://wiki.creativecommons.org/wiki/CC0_FAQ#May_I_apply_CC0_to_computer_software.3F_If_so.2C_is_there_a_recommended_implementation.3F>