Create Deno.gitignore
This is a good file to include for those that are choosing to use the Deno framework for new projects. I believe that the Deno project is used enough to warrant an addition to Github's library of preconfigured .gitignore files.

The contents of the provided file is a direct copy of the .gitignore file provided by the project's own repository. The file I copied is located here: https://github.com/denoland/deno/blob/main/.gitignore

To learn more about the Deno project, please visit https://deno.land

An explanation for the ignored files not mentioned in the comments directly in the file:
```
*.orig # Generally used as a backup when making changes to a file to compare with the original. These can clutter up a repository when they are left as a future reference.
*.pyc # A compiled python script
*.swp # A swap file created by the Vi editor and it's derivatives
.env # A file containing environment variables. Committing this file to a repository can pose a major security risk.

/.cargo_home/ # A directory used by Cargo: https://github.com/rust-lang/cargo#compiling-from-source
/.idea/ # A configuration directory used by JetBrains programs.
/.vscode/ # A configuration directory used by Visual Studio Code
gclient_config.py_entries # I'm not entirely sure what this is beyond that it is related to gclient, see: https://chromium.googlesource.com/chromium/tools/depot_tools.git/+/HEAD/README.gclient.md. If you know more, please feel free to submit a PR with an explanation replacing this comment.
/gh-pages/ # A directory used by Github Pages
/target/ # Often used as an output folder when compiling, building, transpiling, etc. code.
/std/hash/_wasm/target # I assume this is a target folder when compiling web assembly. If you know, please feel free to submit a PR with an explanation replacing this comment.
/tools/wpt/manifest.json # Manifest file for web-platform-tests.
```

Please see the file in the PR for explanations of the rest of the included rules.