Added undo-tree
Added undo-tree, which is a popular package for undoing changes made to a file. It leaves hidden files everywhere, and can litter up a repository if not ignored. 

I have to add `*.~undo-tree~` to every `.gitignore` file in every repository I am working in. As you can guess, this gets very tedious.