fix(go-gitignore): prevent coverage.go files from being excluded
The current `coverage.*` pattern includes `*.go` files as well. If you use `go vendor`, this causes the following files to be excluded:

- `vendor/golang.org/x/text/internal/language/coverage.go`
- `vendor/golang.org/x/text/language/coverage.go`

This becomes a nightmare to troubleshoot why the local build works, yet a CI fails (as the files were not committed).

Adding `!coverage.go` ensure that `*.go` files do not get accidentally excluded.